+++ system/admin/model/adminrolegroupmodel_gen.go
@@ -27,9 +27,9 @@ var (
 type (
 	adminRoleGroupModel interface {
 		Insert(ctx context.Context, data *AdminRoleGroup) (sql.Result, error)
-		FindOne(ctx context.Context, id int64) (*AdminRoleGroup, error)
+		FindOne(ctx context.Context, id int) (*AdminRoleGroup, error)
 		Update(ctx context.Context, data *AdminRoleGroup) error
-		Delete(ctx context.Context, id int64) error
+		Delete(ctx context.Context, id int) error
 	}
 
 	defaultAdminRoleGroupModel struct {
@@ -38,11 +38,11 @@ type (
 	}
 
 	AdminRoleGroup struct {
-		Id       uint `db:"id"`
-		AdminId  uint `db:"admin_id"`
-		RoleId   uint `db:"role_id"`
-		Status   uint8 `db:"status"`    // 状态 0=开启 1=关闭
-		IsDelete uint8 `db:"is_delete"` // 是否删 0=否 1=是
+		Id       int `db:"id"`
+		AdminId  int `db:"admin_id"`
+		RoleId   int `db:"role_id"`
+		Status   int8 `db:"status"`    // 状态 0=开启 1=关闭
+		IsDelete int8 `db:"is_delete"` // 是否删 0=否 1=是
 	}
 )
 
@@ -53,7 +53,7 @@ func newAdminRoleGroupModel(conn sqlx.SqlConn, c cache.CacheConf) *defaultAdminR
 	}
 }
 
-func (m *defaultAdminRoleGroupModel) Delete(ctx context.Context, id int64) error {
+func (m *defaultAdminRoleGroupModel) Delete(ctx context.Context, id int) error {
 	adminRoleGroupIdKey := fmt.Sprintf("%s%v", cacheAdminRoleGroupIdPrefix, id)
 	_, err := m.ExecCtx(ctx, func(ctx context.Context, conn sqlx.SqlConn) (result sql.Result, err error) {
 		query := fmt.Sprintf("delete from %s where `id` = ?", m.table)
@@ -62,7 +62,7 @@ func (m *defaultAdminRoleGroupModel) Delete(ctx context.Context, id int64) error
 	return err
 }
 
-func (m *defaultAdminRoleGroupModel) FindOne(ctx context.Context, id int64) (*AdminRoleGroup, error) {
+func (m *defaultAdminRoleGroupModel) FindOne(ctx context.Context, id int) (*AdminRoleGroup, error) {
 	adminRoleGroupIdKey := fmt.Sprintf("%s%v", cacheAdminRoleGroupIdPrefix, id)
 	var resp AdminRoleGroup
 	err := m.QueryRowCtx(ctx, &resp, adminRoleGroupIdKey, func(ctx context.Context, conn sqlx.SqlConn, v interface{}) error {
